requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
ldap3==2.9.1
pydantic==2.7.1
pydantic-settings==2.2.1
tenacity==8.2.2
prometheus-client==0.20.0
PyYAML==6.0.1
pytest==8.2.2
pytest-asyncio==0.23.6
coverage==7.5.3
requests-mock==1.12.1
pytest-httpx==0.30.0
ruff==0.4.4
black==24.4.2
mypy==1.10.0
